requirements.txt
mysql-connector-python>=8.0.28
scikit-learn>=1.1.1
scipy>=1.7.3
joblib>=1.2.0
numpy>=2.0.0
sentence-transformers>=2.2.2
torch>=2.0.0
tqdm>=4.0.0
